requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.28.1
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.3
uvloop==0.21.0
watchfiles==0.23.0
websockets==15.0.1
pyaudio==0.2.14
librosa==0.11.0
google-genai==1.16.1